chore: pb tutorial feedback
