+++ src/predictions/profiles_mlcorelib/py_native/profiles_tutorial/model.py
@@ -23,6 +23,8 @@ class TutorialModel(BaseModelType):
     }
 
     def __init__(self, build_spec: dict, schema_version: int, pb_version: str) -> None:
+        # FIXME: We should remove this after we introduce a new output type
+        build_spec["materialization"] = {"output_type": "ephemeral"}
         super().__init__(build_spec, schema_version, pb_version)
 
     def get_material_recipe(self) -> PyNativeRecipe:
